versao do dia 17/08

+++ api/src/main/java/com/geargames/api/models/jogos/JogoRepository.java
@@ -1,7 +1,8 @@
 package com.geargames.api.models.jogos;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JogoRepository extends JpaRepository <Jogos,Integer> {
+public interface JogoRepository extends JpaRepository <Jogos, Long> {
     
 }
